Спринт 19.
vol 1.1

1) Созданы HostActivity, navigation_graph и button_navigation_menu.
2) Activity Media, Settings, Search перенесены на фрагменты.
3) Удалены неиспользуемые Activity и xml-файлы верстки.
