+++ app/src/main/AndroidManifest.xml
@@ -17,21 +17,11 @@
         android:theme="@style/Theme.PlaylistMaker.Base.MainMenu"
         tools:targetApi="31">
         <activity
-            android:name=".media.ui.MediaActivity"
-            android:theme="@style/Theme.PlaylistMaker.TabLayout"
-            android:exported="false"/>
-        <activity
-            android:name=".settings.ui.SettingsActivity"
-            android:exported="false"/>
-        <activity
-            android:name=".search.ui.SearchActivity"
-            android:windowSoftInputMode="adjustPan"
-            android:exported="false"/>
-        <activity
             android:name=".audioplayer.ui.AudioPlayerActivity"
             android:exported="false"/>
         <activity
-            android:name=".main.ui.MainActivity"
+            android:name=".main.ui.HostActivity"
+            android:windowSoftInputMode="adjustPan"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
