--- app/src/main/java/com/example/playlistmaker/main/ui/MainActivity.kt
@@ -1,30 +1,0 @@
-package com.example.playlistmaker.main.ui
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.media.ui.MediaActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivityMainBinding
-import com.example.playlistmaker.search.ui.SearchActivity
-import com.example.playlistmaker.settings.ui.SettingsActivity
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val textClickListener: View.OnClickListener = View.OnClickListener {
-            val searchIntent = Intent(this@MainActivity, SearchActivity::class.java)
-            startActivity(searchIntent)
-        }
-        binding.search.setOnClickListener(textClickListener)
-        binding.media.setOnClickListener{ startActivity(Intent(this, MediaActivity::class.java)) }
-        binding.settings.setOnClickListener{ startActivity(Intent(this, SettingsActivity::class.java)) }
-    }
-}
