--- app/src/main/java/com/example/playlistmaker/media/ui/MediaActivity.kt
@@ -1,30 +1,0 @@
-package com.example.playlistmaker.media.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivityMediaBinding
-import com.google.android.material.tabs.TabLayoutMediator
-
-class MediaActivity: AppCompatActivity() {
-    private lateinit var binding: ActivityMediaBinding
-    private lateinit var tabMediator: TabLayoutMediator
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMediaBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.mediaBack.setOnClickListener{ finish() }
-
-        binding.viewPager.adapter = MediaViewPagerAdapter(supportFragmentManager, lifecycle)
-        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
-            tab.text = if (position == 0) getString(R.string.favorites) else getString(R.string.playlists)
-        }
-        tabMediator.attach()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        tabMediator.detach()
-    }
-}
