+++ app/src/main/java/com/example/playlistmaker/search/presentation/view_model/SearchViewModel.kt
@@ -3,7 +3,6 @@ package com.example.playlistmaker.search.presentation.view_model
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.playlistmaker.creator.Creator
 import com.example.playlistmaker.search.domain.api.TrackHistoryInteractor
 import com.example.playlistmaker.search.domain.api.TrackInteractor
 import com.example.playlistmaker.search.domain.models.Track
@@ -16,6 +15,7 @@ class SearchViewModel(
 
     private var _state = MutableLiveData<TrackSearchViewState>()
     val state: LiveData<TrackSearchViewState> get() = _state
+
 
     fun defaultStateValue() {
         _state.value = TrackSearchViewState.Loading
