--- app/src/main/java/com/example/playlistmaker/settings/ui/SettingsActivity.kt
@@ -1,48 +1,0 @@
-package com.example.playlistmaker.settings.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_NO
-import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_YES
-import androidx.appcompat.app.AppCompatDelegate.setDefaultNightMode
-import androidx.lifecycle.ViewModelProvider
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivitySettingsBinding
-import com.example.playlistmaker.settings.presentation.view_model.SettingsViewModel
-import org.koin.androidx.viewmodel.ext.android.viewModel
-
-
-class SettingsActivity: AppCompatActivity()  {
-
-    private lateinit var binding: ActivitySettingsBinding
-    private val settingsViewModel by viewModel<SettingsViewModel>()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_settings)
-
-        binding = ActivitySettingsBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.settingsBack.setOnClickListener{ finish() }
-
-        binding.themeSwitcher.setOnCheckedChangeListener { _, checked ->
-            settingsViewModel.switchTheme(theme = checked)
-        }
-        settingsViewModel.isDarkThemeEnabled.observe(this) {
-            binding.themeSwitcher.isChecked = it
-            applyTheme(it)
-        }
-
-        binding.settingsShare.setOnClickListener{ settingsViewModel.shareApp() }
-        binding.settingsSupport.setOnClickListener{ settingsViewModel.openSupport() }
-        binding.settingsDocs.setOnClickListener{ settingsViewModel.openTerms() }
-
-        settingsViewModel.actionCommand.observe(this) {intent ->
-            intent?.let { startActivity(intent) }
-            settingsViewModel.clearActionCommand()
-        }
-    }
-    private fun applyTheme(isDarkTheme: Boolean) = setDefaultNightMode(if (isDarkTheme) MODE_NIGHT_YES else MODE_NIGHT_NO)
-
-}
