+++ app/src/main/res/drawable/ic_main_menu_media.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="24">
   <path
       android:pathData="M20.5,2H8.5C7.4,2 6.5,2.9 6.5,4V16C6.5,17.1 7.4,18 8.5,18H20.5C21.6,18 22.5,17.1 22.5,16V4C22.5,2.9 21.6,2 20.5,2ZM18.5,7H15.5V12.5C15.5,13.88 14.38,15 13,15C11.62,15 10.5,13.88 10.5,12.5C10.5,11.12 11.62,10 13,10C13.57,10 14.08,10.19 14.5,10.51V5H18.5V7ZM4.5,6H2.5V20C2.5,21.1 3.4,22 4.5,22H18.5V20H4.5V6Z"
-      android:fillColor="#1A1B22"/>
+      android:fillColor="@color/navigation_bar"/>
 </vector>
