+++ app/src/main/res/drawable/ic_small_search.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="16">
     <path
         android:pathData="M12.381,9.805H11.67L11.418,9.589C12.3,8.676 12.832,7.492 12.832,6.203C12.832,3.329 10.211,1 6.978,1C3.746,1 1.125,3.329 1.125,6.203C1.125,9.077 3.746,11.406 6.978,11.406C8.428,11.406 9.761,10.934 10.788,10.149L11.031,10.373V11.006L15.533,15L16.875,13.807L12.381,9.805ZM6.978,9.805C4.736,9.805 2.926,8.196 2.926,6.203C2.926,4.21 4.736,2.601 6.978,2.601C9.221,2.601 11.031,4.21 11.031,6.203C11.031,8.196 9.221,9.805 6.978,9.805Z"
-        android:fillColor="#AEAFB4"/>
+        android:fillColor="@color/navigation_bar"/>
 </vector>
