+++ app/src/main/res/drawable/navigation_view_colored.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="false" android:color="@color/navigation_bar" />
+    <item android:state_checked="true" android:color="@color/light_mode_background" />
+</selector>
