--- app/src/main/res/layout/activity_main.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    tools:context=".main.ui.MainActivity"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="@dimen/padding"
-    android:theme="@style/Theme.PlaylistMaker.Base.MainMenu"
-    android:orientation="vertical">
-
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:text="@string/app_name"
-        android:textSize="22sp">
-    </TextView>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:paddingTop="@dimen/padding">
-
-            <Button
-            android:id="@+id/search"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            style="@style/MainMenuButtonStyle"
-            android:text="@string/main_menu_search"
-            app:icon="@drawable/ic_main_menu_search" />
-
-
-        <Button
-            android:id="@+id/media"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            style="@style/MainMenuButtonStyle"
-            android:text="@string/main_menu_media"
-            app:icon="@drawable/ic_main_menu_media" />
-
-        <Button
-            android:id="@+id/settings"
-            style="@style/MainMenuButtonStyle"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="@string/main_menu_settings"
-            app:icon="@drawable/ic_main_menu_settings" />
-
-    </LinearLayout>
-
-</LinearLayout>
