--- app/src/main/res/layout/activity_media.xml
@@ -1,25 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:app="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-tools:context=".media.ui.MediaActivity"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/mediaBack"
-        style="@style/MediaMenuTextViewStyle">
-    </TextView>
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/tabLayout"
-        style="@style/MediaMenuTabLayoutStyle"
-        app:tabTextAppearance="@style/MediaMenuTabLayoutStyle"
-        app:tabIndicatorFullWidth="false"
-        app:tabMode="fixed"/>
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/viewPager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-</LinearLayout>
