+++ app/src/main/res/layout/fragment_search.xml
@@ -1,0 +1,126 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:tools="http://schemas.android.com/tools"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+android:id="@+id/search"
+tools:context=".search.ui.SearchFragment"
+android:layout_width="match_parent"
+android:layout_height="match_parent"
+android:theme="@style/Theme.PlaylistMaker.Base.Settings"
+android:orientation="vertical">
+
+<TextView
+    android:id="@+id/settingsBack"
+    style="@style/SearchMenuTextViewStyle">
+</TextView>
+
+
+<FrameLayout
+    android:id="@+id/container"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="horizontal"
+    android:padding="@dimen/padding">
+
+    <EditText
+        android:id="@+id/inputEditText"
+        style="@style/EditTextStyle"
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/layout_height_search"
+        android:background="@drawable/edit_text_background"
+        android:drawableStart="@drawable/ic_small_search"
+        android:hint="@string/main_menu_search"
+        android:textCursorDrawable="@drawable/color_cursor"
+        android:imeOptions="actionDone"
+        android:inputType="text"
+        android:maxLength="30"
+        android:maxLines="1"
+        android:paddingStart="@dimen/paddingStart_search"
+        android:singleLine="true" />
+
+    <ImageView
+        android:id="@+id/clearIcon"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end|center"
+        android:background="@drawable/edit_text_background"
+        android:padding="10dp"
+        android:src="@drawable/ic_clear_icon"
+        android:visibility="gone" />
+
+</FrameLayout>
+
+<TextView
+    android:id="@+id/youSearch"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    style="@style/YouSearchTextView"
+    android:visibility="gone">
+</TextView>
+
+<FrameLayout
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <ProgressBar android:id="@+id/progressBar"
+            android:layout_width="44dp"
+            android:layout_height="44dp"
+            android:layout_marginTop="140dp"
+            android:layout_gravity="center_horizontal"
+            android:indeterminateTint="@color/search_progress_bar"
+            android:visibility="gone"/>
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/trackList"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_weight="1"
+            android:layout_marginTop="@dimen/cornerRadius"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            tools:listitem="@layout/track_view" />
+
+        <Button
+            android:id="@+id/clearHistory"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            style="@style/ClearHistoryButton"
+            android:visibility="gone"/>
+
+    </LinearLayout>
+
+    <ImageView
+        android:id="@+id/nothingFound"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        style="@style/NothingFound"
+        android:visibility="gone" />
+
+    <ImageView
+        android:id="@+id/somethingWentWrong"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        style="@style/SomethingWentWrong"
+        android:visibility="gone" />
+
+    <TextView
+        android:id="@+id/errorMessage"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        style="@style/ErrorMessage"
+        android:visibility="gone"
+        tools:text="ERROR" />
+
+    <Button
+        android:id="@+id/refreshButton"
+        android:layout_width="wrap_content"
+        style="@style/RefreshErrorButton"
+        android:visibility="gone"/>
+</FrameLayout>
+
+</LinearLayout>
