+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,17 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/ic_small_search"
+        android:title="@string/main_menu_search" />
+
+    <item
+        android:id="@+id/mediaFragment"
+        android:icon="@drawable/ic_main_menu_media"
+        android:title="@string/main_menu_media" />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/ic_main_menu_settings"
+        android:title="@string/main_menu_settings" />
+</menu>
