+++ app/src/main/res/values-night/colors.xml
@@ -4,4 +4,5 @@
     <color name="pause_icon"> #FFFFFFFF </color>
     <color name="add_and_like_icons"> #AEAFB4 </color>
     <color name="back_icon"> #FFFFFFFF </color>
+    <color name="navigation_bar"> #FFFFFFFF </color>
 </resources>
