+++ app/src/main/res/values-night/styles.xml
@@ -49,12 +49,12 @@
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:gravity">center|start</item>
+        <item name="android:padding">16dp</item>
         <item name="android:text">@string/main_menu_search</item>
         <item name="android:backgroundTint">@color/dark</item>
         <item name="android:textColor">@color/white</item>
         <item name="fontFamily">@font/ys_display_medium</item>
         <item name="android:textSize">22sp</item>
-        <item name="drawableLeftCompat">@drawable/ic_setting_menu_back</item>
         <item name="android:drawableTint">@color/white</item>
     </style>
 
@@ -161,8 +161,8 @@
         <item name="android:textColor">@color/white</item>
         <item name="fontFamily">@font/ys_display_medium</item>
         <item name="android:textSize">22sp</item>
-        <item name="drawableLeftCompat">@drawable/ic_setting_menu_back</item>
         <item name="android:drawableTint">@color/white</item>
+        <item name="android:padding">16dp</item>
     </style>
 
     <style name="MediaMenuTabLayoutStyle">
