+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,6 @@
     <string name="playlists">Плейлисты</string>
     <string name="empty_playlists"> Вы не создали\nни одного плейлиста</string>
     <string name="new_playlist">Новый плейлист</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
